+++ src/aux-common/runtime/AuxCompiler.ts
@@ -7,6 +7,8 @@ import {
 } from './Transpiler';
 import { isFormula, isScript, parseScript, hasValue } from '../bots';
 import { flatMap } from 'lodash';
+import ErrorStackParser from '@casual-simulation/error-stack-parser';
+import StackFrame from 'stackframe';
 
 /**
  * A symbol that identifies a function as having been compiled using the AuxCompiler.
@@ -22,6 +24,135 @@ export class AuxCompiler {
     private _functionCache = new Map<string, Function>();
 
     /**
+     * The offset that should be applied to error line numbers when calculating their original
+     * position. Needed because Node.js Windows produces different line numbers than Mac/Linux.
+     *
+     * Node.js versions greater than v12.14.0 have an issue with identifying the correct line number
+     * for errors and stack traces. This issue is fixed in Node.js v14 and later (possibly also fixed in v13 but I didn't check that).
+     */
+    functionErrorLineOffset: number = 0;
+
+    /**
+     * Calculates the "original" stack trace that the given error occurred at
+     * within the given function.
+     * Returns null if the original stack trace was unable to be determined.
+     * @param functionNameMap A map of function names to their scripts.
+     * @param error The error that occurred.
+     */
+    calculateOriginalStackTrace(
+        functionNameMap: Map<string, AuxCompiledScript>,
+        error: Error
+    ): string {
+        const frames = ErrorStackParser.parse(error);
+
+        if (frames.length < 1) {
+            return null;
+        }
+
+        let transformedFrames: StackFrame[] = [];
+        let lastScriptFrameIndex = -1;
+        let lastScript: AuxCompiledScript;
+        for (let i = frames.length - 1; i >= 0; i--) {
+            const frame = frames[i];
+            const originFrame = frame.evalOrigin;
+            let savedFrame = false;
+            if (
+                !!originFrame &&
+                originFrame.functionName === '__constructFunction'
+            ) {
+                let functionName = frame.functionName;
+                const lastDotIndex = functionName.lastIndexOf('.');
+                if (lastDotIndex >= 0) {
+                    functionName = functionName.slice(lastDotIndex + 1);
+                }
+
+                const script = functionNameMap.get(functionName);
+
+                if (script) {
+                    lastScript = script;
+                    const location: CodeLocation = {
+                        lineNumber:
+                            originFrame.lineNumber +
+                            this.functionErrorLineOffset,
+                        column: originFrame.columnNumber,
+                    };
+                    const originalLocation = this.calculateOriginalLineLocation(
+                        script,
+                        location
+                    );
+                    savedFrame = true;
+                    if (lastScriptFrameIndex < 0) {
+                        lastScriptFrameIndex = i;
+                    }
+                    transformedFrames.unshift(
+                        new StackFrame({
+                            functionName:
+                                lastScript.metadata.diagnosticFunctionName ??
+                                functionName,
+                            fileName: script.metadata.fileName ?? functionName,
+                            lineNumber: originalLocation.lineNumber + 1,
+                            columnNumber: originalLocation.column + 1,
+                        })
+                    );
+                } else if (lastScript) {
+                    const location: CodeLocation = {
+                        lineNumber:
+                            originFrame.lineNumber +
+                            this.functionErrorLineOffset,
+                        column: originFrame.columnNumber,
+                    };
+                    const originalLocation = this.calculateOriginalLineLocation(
+                        lastScript,
+                        location
+                    );
+                    savedFrame = true;
+                    transformedFrames.unshift(
+                        new StackFrame({
+                            functionName:
+                                lastScript.metadata.diagnosticFunctionName ??
+                                functionName,
+                            fileName:
+                                lastScript.metadata.fileName ?? functionName,
+                            lineNumber: originalLocation.lineNumber + 1,
+                            columnNumber: originalLocation.column + 1,
+                        })
+                    );
+                }
+            }
+
+            if (!savedFrame) {
+                if (frame.functionName === '__wrapperFunc') {
+                    savedFrame = true;
+                    if (lastScriptFrameIndex > i) {
+                        lastScriptFrameIndex -= 1;
+                    }
+                }
+            }
+
+            if (!savedFrame) {
+                transformedFrames.unshift(frame);
+            }
+        }
+
+        if (lastScriptFrameIndex >= 0) {
+            const finalFrames = [
+                ...transformedFrames.slice(0, lastScriptFrameIndex + 1),
+                new StackFrame({
+                    fileName: '[Native CasualOS Code]',
+                    functionName: '<CasualOS>',
+                }),
+            ];
+
+            const stack = finalFrames
+                .map((frame) => '   at ' + frame.toString())
+                .join('\n');
+            return error.toString() + '\n' + stack;
+        }
+
+        return null;
+    }
+
+    /**
      * Calculates the original location within the given function for the given location.
      * @param func The function.
      * @param location The location.
@@ -30,8 +161,8 @@ export class AuxCompiler {
         func: AuxCompiledScript,
         location: CodeLocation
     ): CodeLocation {
-        // Line numbers should be zero based
-        if (location.lineNumber < func.metadata.scriptLineOffset - 1) {
+        // Line numbers should be one based
+        if (location.lineNumber < func.metadata.scriptLineOffset) {
             return {
                 lineNumber: 0,
                 column: 0,
@@ -39,8 +170,9 @@ export class AuxCompiler {
         }
 
         let transpiledLocation: CodeLocation = {
-            lineNumber: location.lineNumber - func.metadata.scriptLineOffset,
-            column: location.column,
+            lineNumber:
+                location.lineNumber - func.metadata.scriptLineOffset - 1,
+            column: location.column - 1,
         };
 
         return calculateOriginalLineLocation(
@@ -70,6 +202,9 @@ export class AuxCompiler {
             scriptFunction,
             scriptLineOffset,
             transpilerResult,
+            fileName: options?.fileName,
+            diagnosticFunctionName: options?.diagnosticFunctionName,
+            isAsync: async,
         };
 
         if (options) {
@@ -96,16 +231,18 @@ export class AuxCompiler {
                           invoke(() => scriptFunc(...args), context)
                     : scriptFunc;
                 if (async) {
-                    func = function (...args: any[]) {
+                    func = function __wrapperFunc(...args: any[]) {
                         before(context);
                         try {
-                            const result = finalFunc(...args);
+                            let result = finalFunc(...args);
                             if (!(result instanceof Promise)) {
-                                return new Promise((resolve, reject) => {
+                                result = new Promise((resolve, reject) => {
                                     result.then(resolve, reject);
                                 });
                             }
-                            return result;
+                            return result.catch((ex: any) => {
+                                onError(ex, context, meta);
+                            });
                         } catch (ex) {
                             onError(ex, context, meta);
                         } finally {
@@ -113,7 +250,7 @@ export class AuxCompiler {
                         }
                     };
                 } else {
-                    func = function (...args: any[]) {
+                    func = function __wrapperFunc(...args: any[]) {
                         before(context);
                         try {
                             return finalFunc(...args);
@@ -194,6 +331,7 @@ export class AuxCompiler {
         }
         script = this._parseScript(script);
         let scriptLineOffset = 0;
+        let constantsLineOffset = 0;
 
         let constantsCode = '';
         if (options.constants) {
@@ -201,7 +339,7 @@ export class AuxCompiler {
                 .filter((v) => v !== 'this')
                 .map((v) => `const ${v} = constants["${v}"];`);
             constantsCode = lines.join('\n') + '\n';
-            scriptLineOffset += 1 + Math.max(lines.length - 1, 0);
+            constantsLineOffset += 1 + Math.max(lines.length - 1, 0);
         }
 
         let variablesCode = '';
@@ -235,44 +373,73 @@ export class AuxCompiler {
 
         // Function needs a name because acorn doesn't understand
         // that this function is allowed to be anonymous.
-        let functionCode = `function _(...args) { ${argumentsCode}${variablesCode}${scriptCode}\n }`;
+        let functionCode = `function ${
+            options.functionName ?? '_'
+        }(...args) { ${argumentsCode}${variablesCode}${scriptCode}\n }`;
         if (async) {
             functionCode = `async ` + functionCode;
         }
-        const transpiled = this._transpiler.transpileWithMetadata(functionCode);
+        try {
+            const transpiled = this._transpiler.transpileWithMetadata(
+                functionCode
+            );
 
-        const finalCode = `${constantsCode}return ${transpiled.code};`;
+            scriptLineOffset += constantsLineOffset;
+            const finalCode = `${constantsCode}return ${transpiled.code};`;
 
-        let func = this._buildFunction(finalCode, options);
-        (<any>func)[COMPILED_SCRIPT_SYMBOL] = true;
+            let func = this._buildFunction(finalCode, options);
+            (<any>func)[COMPILED_SCRIPT_SYMBOL] = true;
 
-        // Add 1 extra line to count the line feeds that
-        // is automatically inserted at the start of the script as part of the process of
-        // compiling the dynamic script.
-        // See https://tc39.es/ecma262/#sec-createdynamicfunction
-        scriptLineOffset += 1;
+            // Add 1 extra line to count the line feeds that
+            // is automatically inserted at the start of the script as part of the process of
+            // compiling the dynamic script.
+            // See https://tc39.es/ecma262/#sec-createdynamicfunction
+            scriptLineOffset += 2;
 
-        if (options.variables) {
-            if ('this' in options.variables) {
-                func = func.bind(options.variables['this'](options.context));
+            if (options.variables) {
+                if ('this' in options.variables) {
+                    func = func.bind(
+                        options.variables['this'](options.context)
+                    );
+                }
             }
-        }
 
-        return { func, scriptLineOffset, async, transpilerResult: transpiled };
+            return {
+                func,
+                scriptLineOffset,
+                async,
+                transpilerResult: transpiled,
+            };
+        } catch (err) {
+            if (err instanceof SyntaxError) {
+                const replaced = replaceSyntaxErrorLineNumber(
+                    err,
+                    (location) => ({
+                        lineNumber: location.lineNumber - scriptLineOffset,
+                        column: location.column,
+                    })
+                );
+
+                if (replaced) {
+                    throw replaced;
+                }
+            }
+            throw err;
+        }
     }
 
     private _buildFunction<T>(
         finalCode: string,
         options: AuxCompileOptions<T>
     ) {
-        return this._constructFunction<T>(finalCode)(
+        return this.__constructFunction<T>(finalCode)(
             options.constants,
             options.variables,
             options.context
         );
     }
 
-    private _constructFunction<T>(
+    private __constructFunction<T>(
         finalCode: string
     ): (
         constants: AuxCompileOptions<T>['constants'],
@@ -287,9 +454,47 @@ export class AuxCompiler {
                 'context',
                 finalCode
             ) as any;
+
             this._functionCache.set(finalCode, existing);
         }
         return existing;
+    }
+}
+
+const SYNTAX_ERROR_LINE_NUMBER_REGEX = /\((\d+):(\d+)\)$/;
+
+/**
+ * Parses the line and column numbers from the the given syntax error, transforms them with the given function,
+ * and returns a new syntax error that contains the new location. Returns null if the line and column numbers could not be parsed.
+ * @param error The error to transform.
+ * @param transform The function that should be used to transform the errors.
+ * @returns
+ */
+export function replaceSyntaxErrorLineNumber(
+    error: SyntaxError,
+    transform: (location: CodeLocation) => CodeLocation
+): SyntaxError {
+    const matches = SYNTAX_ERROR_LINE_NUMBER_REGEX.exec(error.message);
+
+    if (matches) {
+        const [str, line, column] = matches;
+
+        const lineNumber = parseInt(line);
+        const columnNumber = parseInt(column);
+
+        const location = transform({
+            lineNumber,
+            column: columnNumber,
+        });
+
+        return new SyntaxError(
+            error.message.replace(
+                str,
+                `(${location.lineNumber}:${location.column})`
+            )
+        );
+    } else {
+        return null;
     }
 }
 
@@ -338,6 +543,21 @@ export interface AuxScriptMetadata {
      * The transpiler result;
      */
     transpilerResult: TranspilerResult;
+
+    /**
+     * The name that should be used for the function in stack traces.
+     */
+    diagnosticFunctionName: string;
+
+    /**
+     * The file name that was specified for the script.
+     */
+    fileName: string;
+
+    /**
+     * Whether the function is asynchronous and returns a promise.
+     */
+    isAsync: boolean;
 }
 
 /**
@@ -387,6 +607,21 @@ export interface AuxCompileOptions<T> {
      * A function that should be called when an error occurs.
      */
     onError?: (error: any, context?: T, meta?: AuxScriptMetadata) => void;
+
+    /**
+     * The name that should be given to the function.
+     */
+    functionName?: string;
+
+    /**
+     * The name that should be used for the function in stack traces.
+     */
+    diagnosticFunctionName?: string;
+
+    /**
+     * The file name that should be used for transformed error stack traces.
+     */
+    fileName?: string;
 }
 
 // export class CompiledScriptError extends Error {